SmlAdminDataCalculator moved to uk.ons.registers.methods
